requirements.txt
pillow>=10.0.0
opencv-python>=4.8.0
pytesseract>=0.3.10
langdetect>=1.0.9
numpy>=1.24.0
pytest>=7.4.0

